+++ app/linux/flutter/generated_plugins.cmake
@@ -5,6 +5,7 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   desktop_drop
   file_selector_linux
+  flutter_js
   irondash_engine_context
   media_kit_libs_linux
   media_kit_video